[Fix] When converting instruments to samples and there are more than 255 samples, do not write possibly bogus sample numbers into the patterns. [Ref] Silence two minor 64-bit warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5658 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/View_tre.cpp
@@ -335,7 +335,7 @@ BOOL CModTree::PreTranslateMessage(MSG *pMsg)
 		CInputHandler* ih = (CMainFrame::GetMainFrame())->GetInputHandler();
 
 		//Translate message manually
-		UINT nChar = pMsg->wParam;
+		UINT nChar = (UINT)pMsg->wParam;
 		UINT nRepCnt = LOWORD(pMsg->lParam);
 		UINT nFlags = HIWORD(pMsg->lParam);
 		KeyEventType kT = ih->GetKeyEventType(nFlags);
